refactors class Map in new file
