+++ ggjxx/src/TerminalUI/tui.py
@@ -2,7 +2,8 @@
 # =================== DEPRECATED ===================
 # ==================================================
 
-from ..game.game import Game, Map, Robot, FireFighter, RadiationFighter
+from ..game.game import Game, Robot, FireFighter, RadiationFighter
+from ..game.map import Map
 from ..game.constants import *
 from ..input.getch import _Getch
